+++ Lib/Bundler/IBundlerCallback.cs
@@ -7,6 +7,7 @@ namespace Lib.Bundler
     public interface IBundlerCallback
     {
         string ReadContent(string name);
+        IList<string> GetPlainJsDependencies(string name);
         void WriteBundle(string name, string content);
         /// <summary>
         /// Empty forName means main bundle.js