+++ Lib/Bundler/Js/Tests/Expected/SimpleWithJsDep/cb-bundle.js
@@ -1,0 +1,7 @@
+console.log("I am dependency"), function(undefined) {
+    "use strict";
+    function hello() {
+        return "Hello";
+    }
+    console.log(hello());
+}();