+++ Lib/Bundler/Js/Tests/Expected/SimpleWithJsDep/cbm-bundle.js
@@ -1,0 +1,7 @@
+console.log("I am dependency"), function(n) {
+    "use strict";
+    function o() {
+        return "Hello";
+    }
+    console.log(o());
+}();