+++ Lib/Bundler/Js/Tests/Expected/SimpleWithJsDep/cm-bundle.js
@@ -1,0 +1,1 @@
+console.log("I am dependency"),function(n){"use strict";function o(){return"Hello"}console.log(o())}();