+++ Lib/Bundler/Js/Tests/Inputs/SimpleWithJsDep/lib.js
@@ -1,0 +1,5 @@
+function hello() {
+    return "Hello";
+}
+
+exports.hello = hello;