+++ Lib/Bundler/Js/bundler.ts
@@ -4,6 +4,8 @@ declare const bb: IBB;
 
 interface IBB {
     readContent(name: string): string;
+    /// returns pipe delimited list of file names
+    getPlainJsDependencies(name: string): string;
     writeBundle(name: string, content: string): void;
     generateBundleName(forName: string): string;
     resolveRequire(name: string, from: string): string;
@@ -23,6 +25,7 @@ interface IFileForBundle {
     selfexports: { name?: string; node?: IAstNode; reexport?: string }[];
     exports?: { [name: string]: IAstNode };
     pureFuncs: { [name: string]: boolean };
+    plainJsDependencies: string[];
 }
 
 type FileForBundleCache = { [name: string]: IFileForBundle };
@@ -227,6 +230,8 @@ function check(
     let ast = parse(fileContent);
     //console.log(ast.print_to_string({ beautify: true }));
     ast.figure_out_scope!();
+    let jsdeps = bb.getPlainJsDependencies(name).split("|");
+    if (jsdeps.length == 1 && jsdeps[0] == "") jsdeps = [];
     cached = {
         name,
         ast,
@@ -236,7 +241,8 @@ function check(
         partOfBundle: requiredAs,
         selfexports: [],
         exports: undefined,
-        pureFuncs: Object.create(null)
+        pureFuncs: Object.create(null),
+        plainJsDependencies: jsdeps
     };
     const cachedName = name.toLowerCase();
     cache[cachedName] = cached;
@@ -272,8 +278,8 @@ __bbe['${cached.shortname}']=module.exports; }).call(window);`);
     let reexportDef: ISymbolDef | undefined;
     let walker = new TreeWalker((node: IAstNode, descend: () => void) => {
         if (node instanceof AST_Block) {
-            (<IAstBlock>node).body = (<IAstBlock>node).body!
-                .map((stm): IAstStatement | undefined => {
+            (<IAstBlock>node).body = (<IAstBlock>node)
+                .body!.map((stm): IAstStatement | undefined => {
                     if (stm instanceof AST_Directive && stm.value === "use strict") {
                         return undefined;
                     } else if (stm instanceof AST_SimpleStatement) {
@@ -530,6 +536,7 @@ interface ISplitInfo {
     exportsAllUsedFromLazyBundles: Map<string, string>;
     directSplitsForcedLazy: Set<ISplitInfo>;
     expandedSplitsForcedLazy: ISplitInfo[];
+    plainJsDependencies: Set<string>;
 }
 
 type SplitMap = { [name: string]: ISplitInfo };
@@ -578,7 +585,8 @@ function bundle(project: IBundleProject) {
         importsFromOtherBundles: new Map(),
         exportsAllUsedFromLazyBundles: new Map(),
         directSplitsForcedLazy: new Set(),
-        expandedSplitsForcedLazy: []
+        expandedSplitsForcedLazy: [],
+        plainJsDependencies: new Set()
     };
     let lastGeneratedIdentId = 0;
     function generateIdent(): string {
@@ -586,19 +594,23 @@ function bundle(project: IBundleProject) {
     }
     order.forEach(f => {
         let fullBundleName = f.partOfBundle;
-        if (bundleNames.indexOf(fullBundleName) >= 0) return;
-        let shortenedBundleName = bb.generateBundleName(fullBundleName);
-        bundleNames.push(fullBundleName);
-        splitMap[fullBundleName] = {
-            fullName: fullBundleName,
-            shortName: shortenedBundleName,
-            propName: generateIdent(),
-            exportsUsedFromLazyBundles: new Map(),
-            importsFromOtherBundles: new Map(),
-            exportsAllUsedFromLazyBundles: new Map(),
-            directSplitsForcedLazy: new Set(),
-            expandedSplitsForcedLazy: []
-        };
+        if (bundleNames.indexOf(fullBundleName) < 0) {
+            let shortenedBundleName = bb.generateBundleName(fullBundleName);
+            bundleNames.push(fullBundleName);
+            splitMap[fullBundleName] = {
+                fullName: fullBundleName,
+                shortName: shortenedBundleName,
+                propName: generateIdent(),
+                exportsUsedFromLazyBundles: new Map(),
+                importsFromOtherBundles: new Map(),
+                exportsAllUsedFromLazyBundles: new Map(),
+                directSplitsForcedLazy: new Set(),
+                expandedSplitsForcedLazy: [],
+                plainJsDependencies: new Set()
+            };
+        }
+        let jsDeps = splitMap[fullBundleName].plainJsDependencies;
+        f.plainJsDependencies.forEach(v => jsDeps.add(v));
     });
     if (bundleNames.length > 1) detectBundleExportsImports(order, splitMap, cache, generateIdent);
     for (let bundleIndex = 0; bundleIndex < bundleNames.length; bundleIndex++) {
@@ -765,7 +777,16 @@ function bundle(project: IBundleProject) {
             );
             f.ast.transform!(transformer);
         });
-        appendExportedFromLazyBundle(bodyAst, splitMap[currentBundleName], cache);
+        let currentSplit = splitMap[currentBundleName];
+        appendExportedFromLazyBundle(bodyAst, currentSplit, cache);
+        let jsDepAsts: IAstStatement[] = [];
+        for (const jsDep of currentSplit.plainJsDependencies.keys()) {
+            let depAst = parse(bb.readContent(jsDep));
+            jsDepAsts = jsDepAsts.concat(depAst.body!);
+        }
+        if (jsDepAsts.length > 0) {
+            bundleAst.body!.unshift(...jsDepAsts);
+        }
         bundleAst = compressAst(project, bundleAst, pureFuncs);
         mangleNames(project, bundleAst);
         let out = printAst(project, bundleAst);